move all matrix stuff
