+++ poseidon-parameters/Cargo.toml
@@ -12,3 +12,11 @@ anyhow = { version = "1", default-features = false }
 ark-ff = { version = "^0.3.0", default-features = false }
 ark-std = { version = "^0.3.0", default-features = false }
 num-integer = { version = "0.1.45", default-features = false }
+
+[dev-dependencies]
+ark-bls12-377 = "0.3"
+ark-bls12-381 = "0.3"
+ark-bn254 = "0.3"
+ark-ed-on-bls12-377 = "0.3"
+ark-ed-on-bls12-381 = "0.3"
+proptest = "1"
